requirements.txt
Flask
SQLAlchemy
discord.py==2.3.2
flask
gunicorn
sqlalchemy==1.4.41
werkzeug